plots paired-end data for a single chromosome

plots paired-end data for a single chromosome

bb_plotBedpe(
  bedpe,
  chrom,
  params = NULL,
  chromstart = NULL,
  chromend = NULL,
  fillcolor = "black",
  colorby = NULL,
  colorbyrange = NULL,
  linecolor = NA,
  assembly = "hg19",
  boxHeight = unit(2, "mm"),
  spaceHeight = 0.3,
  spaceWidth = 0.02,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = c("left", "top"),
  default.units = "inches",
  draw = TRUE,
  ...
)

Arguments

bedpe
bed paired end data to be plotted

chrom
chromsome of region to be plotted

params
an optional "bb_params" object space containing relevant function parameters

chromstart
start position

chromend
end position

fillcolor
single value, vector, or palette specifying colors of bedpe elements

colorby
name of column in bedpe data to scale fillcolors by

colorbyrange
the range of values to apply a colorby palette scale to, if colorby values are numeric

linecolor
border color

assembly
default genome assembly as a string or a bb_assembly object

boxHeight
A numeric or unit object specifying height of boxes at either end of bedpe element

spaceHeight
height of space between boxes of different bedpe elements

spaceWidth
width of spacing between bedpe elements, as a fraction of the plot's genomic range

x
A numeric vector or unit object specifying x-location

y
A numeric vector or unit object specifying y-location

width
A numeric vector or unit object specifying width

height
A numeric vector or unit object specifying height

just
string or numeric vector specifying the justification of the viewport relative to its (x, y) location: "left", "right", "centre", "center", "bottom", "top"

default.units
A string indicating the default units to use if x, y, width, or height are only given as numeric vectors

draw
A logical value indicating whether graphics output should be produced