Plots genomic coordinates for the axis of a BentoBox plot

Plots genomic coordinates for the axis of a BentoBox plot

bb_plotGenomeLabel(
  x,
  y,
  plot = NULL,
  width = NULL,
  chrom = NULL,
  chromstart = NULL,
  chromend = NULL,
  params = NULL,
  just = c("left", "top"),
  scale = "bp",
  assembly = "hg19",
  fontsize = 10,
  fontcolor = "black",
  linecolor = "black",
  commas = TRUE,
  sequence = TRUE,
  boxWidth = 0.5,
  ticks = NULL,
  tcl = 0.5,
  default.units = "inches",
  ...
)

Arguments

x
A numeric or unit object specifying x-location.

y
A numeric or unit object specifying y-location.

plot
BentoBox plot to add genome label

width
if not specifying plot, width of genome label, as a numeric or unit object

chrom
if not specifying plot, chromosome of genome label

chromstart
if not specifying plot, chromstart of genome label

chromend
if not specifying plot, chromend of genome label

params
an optional "bb_params" object space containing relevant function parameters

just
The justification of the label relative to its (x, y) location.If there are two values, the first specifies horizontal justification and the second value specifies vertical justification. Possible string values are: "left", "right", "centre", "center", "bottom", and "top".

scale
scale of the genome of the label. Options are "bp", "Kb", or "Mb"

assembly
default genome assembly as a string or a bb_assembly object (will be inherited from plot object if provided)

fontsize
Fontsize for labels (in points).

fontcolor
Color for all text and lines (with the exception of sequence information).

linecolor
Axis linecolor.

commas
A logical value indicating whether to include commas in start and stop labels.

sequence
A logical value indicating whether to include sequence information above the label axis (only at appropriate resolutions).

ticks
A numeric vector of x-value locations for tick marks.

tcl
Length of tickmark as fraction of text height.

default.units
A string indicating the default units to use if x or y are only given as numerics.

lwd
Axis linewidth.

fontfamily
Text fontfamily.