plots a Manhattan plot

plots a Manhattan plot

bb_plotManhattan(
  bed,
  pVals,
  params = NULL,
  chrom = NULL,
  chromstart = NULL,
  chromend = NULL,
  assembly = "hg19",
  fillcolor = "black",
  pch = 19,
  space = 0.01,
  cex = 0.25,
  ymax = 1,
  range = NULL,
  sigVal = 5e-08,
  sigLine = F,
  sigCol = NULL,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = c("left", "top"),
  default.units = "inches",
  draw = T,
  ...
)

Arguments

bed
data for Manhattan plot, can be any data in BED format or a GRanges

pVals
name of column in bedfile of corresponding p-values (will be converted to -log(10) space)

params
an optional "bb_params" object space containing relevant function parameters

chrom
string of specific chromosome to zoom in on according to genome build (Ex: hg19 = "chr1"); if NULL all chromosomes found in data will be plotted

chromstart
if single chromosome specified, start of region to be plotted

chromend
if single chromosome specified, end of region to be plotted

assembly
default genome assembly as a string or a bb_assembly object

space
the space between each chromsome as a fraction of the width of the plot

cex
number indiciating the amount by which points should be scaled relative to the default

ymax
fraction of max y value to set as height of plot

range
y-range of p-values to plot (c(min, max))

sigVal
numeric indicating the significance level

sigLine
logical indicating whether to draw a line at the significance level

sigCol
single color for coloring significant values

x
A numeric or unit object specifying x-location

y
A numeric or unit object specifying y-location

width
A numeric or unit object specifying width

height
A numeric or unit object specifying height

just
A string or numeric vector specifying the justification of the viewport relative to its (x, y) location: "left", "right", "centre", "center", "bottom", "top"

default.units
A string indicating the default units to use if x, y, width, or height are only given as numerics

draw
A logical value indicating whether graphics output should be produced

colors
single color, vector of colors, or color palette

transp
transparency of colors