plots gene transcripts in a pileup style

plots gene transcripts in a pileup style

bb_plotTranscripts(
  chrom,
  params = NULL,
  assembly = "hg19",
  chromstart = NULL,
  chromend = NULL,
  boxHeight = unit(2, "mm"),
  spaceHeight = 0.3,
  spaceWidth = 0.02,
  fillcolor = c("#8a8aff", "#ff7e7e"),
  colorbyStrand = TRUE,
  labels = "transcript",
  fontsize = 8,
  strandSplit = FALSE,
  stroke = 0.1,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = c("left", "top"),
  default.units = "inches",
  draw = TRUE
)

Arguments

chrom
chromsome of region to be plotted

params
an optional "bb_params" object space containing relevant function parameters

assembly
desired genome assembly

chromstart
start position

chromend
end position

boxHeight
A numeric or unit object specifying total height of each transcript

spaceHeight
height of spacing between transcripts, as a fraction of boxHeight

spaceWidth
width of minimum spacing between transcripts, as a fraction of the plot's genomic range

fillcolor
single value or vector specifying colors of transcripts

colorbyStrand
a logical value indicating whether to color strands by the first two colors in a fillcolor vector

labels
how labels should be determined; options are NULL for no labels, "transcript" for transcript names, "gene" for gene names, or "both" for combined transcript and gene names

fontsize
the size of gene label text (in points)

strandSplit
logical indicating whether plus and minus-stranded elements should be separated

stroke
numerical value indicating the stroke width for transcript body outlines

x
A numeric or unit object specifying x-location

y
A numeric or unit object specifying y-location

width
A numeric or unit object specifying width

height
A numeric or unit object specifying height

just
string or numeric vector specifying the justification of the viewport relative to its (x, y) location: "left", "right", "centre", "center", "bottom", "top"

default.units
A string indicating the default units to use if x, y, width, or height are only given as numerics

draw
A logical value indicating whether graphics output should be produced

Value

Function will plot a pileup of transcripts and return a bb_transcripts object